public_site/requirements.txt
asgiref==3.6.0
azure-common==1.1.28
azure-core==1.28.0
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==3.2.18
django-environ==0.10.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.7.0
isodate==0.6.1
msal==1.22.0
msal-extensions==1.0.0
packaging==23.1
Pillow==9.5.0
pkg_resources==0.0.0
portalocker==2.7.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==2.0.3
zipp==3.15.0
